pyproject.toml
[project]
name = "etl-sandbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "snowflake-connector-python>=3.17.3",
]
